Linux helper cleanup
