+++ chipsec/helper/linux/helper.py
@@ -51,8 +51,6 @@ from chipsec.helper.basehelper import Helper
 from chipsec.logger import logger, print_buffer
 import chipsec.file
 
-from chipsec_tools import efi_compressor
-
 MSGBUS_MDR_IN_MASK  = 0x1
 MSGBUS_MDR_OUT_MASK = 0x2
 
@@ -79,9 +77,10 @@ IOCTL_VA2PA                    = 0x14
 IOCTL_MSGBUS_SEND_MESSAGE      = 0x15
 IOCTL_FREE_PHYSMEM             = 0x16
 
-LZMA  = efi_compressor.LzmaDecompress
-Tiano = efi_compressor.TianoDecompress
-EFI   = efi_compressor.EfiDecompress
+LZMA  = os.path.join(chipsec.file.TOOLS_DIR,"compression","bin","LzmaCompress")
+TIANO = os.path.join(chipsec.file.TOOLS_DIR,"compression","bin","TianoCompress")
+EFI   = os.path.join(chipsec.file.TOOLS_DIR,"compression","bin","TianoCompress") + "-uefi"
+BROTLI = os.path.join(chipsec.file.TOOLS_DIR,"compression","bin","Brotli")
 
 class MemoryMapping(mmap.mmap):
     """Memory mapping based on Python's mmap.
@@ -104,8 +103,8 @@ class LinuxHelper(Helper):
     SUPPORT_KERNEL26_GET_PHYS_MEM_ACCESS_PROT = False
     DKMS_DIR = "/var/lib/dkms/"
 
-    decompression_oder_type1 = [Tiano, EFI]
-    decompression_oder_type2 = [LZMA, Tiano, EFI]
+    decompression_oder_type1 = [chipsec.defines.COMPRESSION_TYPE_TIANO,chipsec.defines.COMPRESSION_TYPE_UEFI]
+    decompression_oder_type2 = [chipsec.defines.COMPRESSION_TYPE_TIANO,chipsec.defines.COMPRESSION_TYPE_UEFI,chipsec.defines.COMPRESSION_TYPE_LZMA,chipsec.defines.COMPRESSION_TYPE_BROTLI]
 
     def __init__(self):
         super(LinuxHelper, self).__init__()
@@ -1059,50 +1058,81 @@ class LinuxHelper(Helper):
     def rotate_list(self, list, n):
         return list[n:] + list[:n]
 
-    def rotate_algorithms(self, CompressionType, n):
-        if(CompressionType == 0x01):
-            self.decompression_oder_type1 = self.rotate_list(self.decompression_oder_type1, n)
-        elif(CompressionType == 0x02):
-            self.decompression_oder_type2 = self.rotate_list(self.decompression_oder_type2, n)
-
-    def decompress_data(self, funcs, cdata, CompressionType):
+    def unknown_decompress(self,CompressedFileName,OutputFileName):
         failed_times = 0
-        for func in funcs:
-            try:
-                data = func(cdata, len(cdata))
-                if(failed_times > 0):
-                    self.rotate_algorithms(CompressionType, failed_times)
-                return data
-            except Exception:
+        for CompressionType in [self.decompression_oder_type2]:
+            res = self.decompress_file(CompressedFileName,OutputFileName,CompressionType)
+            if res == True:
+                self.rotate_list(self.decompression_oder_type2,failed_times)
+                break
+            else:
                 failed_times += 1
-                continue
-        return None
+        return res
+        
+    def unknown_efi_decompress(self,CompressedFileName,OutputFileName):
+        failed_times = 0
+        for CompressionType in [self.decompression_oder_type1]:
+            res = self.decompress_file(CompressedFileName,OutputFileName,CompressionType)
+            if res == True:
+                self.rotate_list(self.decompression_oder_type1,failed_times)
+                break
+            else:
+                failed_times += 1
+        return res
+
     #
-    # Decompress binary with efi_compressor from https://github.com/theopolis/uefi-firmware-parser
+    # Compress binary file
+    #
+    def compress_file( self, FileName, OutputFileName, CompressionType ):
+        if not CompressionType in [i for i in chipsec.defines.COMPRESSION_TYPES]:
+            return False
+        encode_str = " -e -o {} {}".format(OutputFileName,FileName)
+        if CompressionType == chipsec.defines.COMPRESSION_TYPE_NONE:
+            shutil.copyfile(FileName,OutputFileName)
+            return True
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_TIANO:
+            encode_str = TIANO + encode_str
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_UEFI:
+            encode_str = EFI + encode_str
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_LZMA:
+            encode_str = LZMA + encode_str
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_BROTLI:
+            encode_str = BROTLI + encode_str
+        data = subprocess.check_output(encode_str,shell=True)
+        if not data == 0 and logger().VERBOSE:
+            logger().error("Cannot decompress file({})".format(CompressedFileName))
+            return False
+        return True
+        
+    #
+    # Decompress binary
     #
     def decompress_file( self, CompressedFileName, OutputFileName, CompressionType ):
-        CompressedFileData = chipsec.file.read_file( CompressedFileName )
-        if CompressionType == 0: # not compressed
-            shutil.copyfile( CompressedFileName, OutputFileName )
-        elif CompressionType == 0x01:
-            data = self.decompress_data( self.decompression_oder_type1, CompressedFileData, CompressionType)
-        elif CompressionType == 0x02:
-            data = self.decompress_data( self.decompression_oder_type2, CompressedFileData, CompressionType)
-        if CompressionType != 0x00:
-            if data is not None:
-                chipsec.file.write_file( OutputFileName, data )
-            else:
-                if len(CompressedFileData) > 4:
-                    data = self.decompress_data( self.decompression_oder_type2, CompressedFileData[4:], CompressionType )
-                    if data is not None:
-                        chipsec.file.write_file( OutputFileName, data )
-                    else:
-                        if logger().DEBUG: logger().error( "Cannot decompress file (%s)" % ( CompressedFileName ) )
-                        return None
-                else:
-                    if logger().DEBUG: logger().error( "Cannot decompress file (%s) with Data size (%d) " % ( CompressedFileName, len(CompressedFileData) ) )
-                    return None
-        return chipsec.file.read_file( OutputFileName )
+        if not CompressionType in [i for i in chipsec.defines.COMPRESSION_TYPES]:
+            return False
+        if CompressionType == chipsec.defines.COMPRESSION_TYPE_UNKNOWN:
+            data = self.unknown_decompress(CompressedFileName,OutputFileName)
+            return data
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_EFI_STANDARD:
+            data = self.unknown_efi_decompress(CompressedFileName,OutputFileName)
+            return data
+        decode_str = " -d -o {} {}".format(OutputFileName, CompressedFileName)
+        if CompressionType == chipsec.defines.COMPRESSION_TYPE_NONE:
+            shutil.copyfile(CompressedFileName,OutputFileName)
+            return True
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_TIANO:
+            decode_str = TIANO + decode_str
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_UEFI:
+            decode_str = EFI + decode_str
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_LZMA:
+            decode_str = LZMA + decode_str
+        elif CompressionType == chipsec.defines.COMPRESSION_TYPE_BROTLI:
+            decode_str = BROTLI + decode_str
+        data = subprocess.check_output(decode_str,shell=True)
+        if not data == 0 and logger().VERBOSE:
+            logger().error("Cannot decompress file({})".format(CompressedFileName))
+            return False
+        return True
 
     #
     # Logical CPU count
